+++ src/prefect/__init__.py
@@ -12,7 +12,7 @@ from prefect.core import Task, Flow, Parameter
 import prefect.engine
 import prefect.tasks
 from prefect.tasks.control_flow import case
-from prefect.tasks.resources import resource_manager
+from prefect.tasks.core.resource_manager import resource_manager
 from prefect.utilities.tasks import task, tags, unmapped, apply_map
 
 import prefect.serialization
