+++ src/prefect/tasks/core/__init__.py
@@ -6,3 +6,4 @@ import prefect.tasks.core.collections
 import prefect.tasks.core.constants
 import prefect.tasks.core.function
 import prefect.tasks.core.operators
+import prefect.tasks.core.resource_manager
