--- src/prefect/tasks/resources/__init__.py
@@ -1,1 +1,0 @@
-from .base import resource_manager, ResourceManager
